Separate admin subroutes into their own components
