+++ src/routes/admin/Reports.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Header } from '../../components/Typography';
+
+
+function Reports() {
+    return <>
+        <Header>Reports</Header>
+    </>;
+}
+
+export default Reports;
