+++ src/routes/admin/Users.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Header } from '../../components/Typography';
+
+
+function Users() {
+    return <>
+        <Header>Users</Header>
+    </>;
+}
+
+export default Users;
